+++ src/invest/common/invest-contract-card/index.ts
@@ -1,0 +1,1 @@
+export * from './invest-contract-card'
